requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
cellpose==3.1.1.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastremap==1.15.0
filelock==3.18.0
fill_voids==2.1.0
fire==0.7.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
imagecodecs==2025.3.30
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.5
monai==1.5.0
mpmath==1.3.0
narwhals==2.0.1
natsort==8.4.0
networkx==3.4.2
numba==0.61.2
numpy==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.2
pytz==2025.2
referencing==0.36.2
requests==2.32.4
roifile==2024.5.24
rpds-py==0.26.0
scipy==1.15.3
segment-anything==1.0
six==1.17.0
smmap==5.0.2
streamlit==1.47.1
sympy==1.13.1
tenacity==9.1.2
termcolor==3.1.0
tifffile==2025.5.10
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
